feat: TattooCard 컴포넌트 분리
